# openapi/openapi.yaml
openapi: 3.1.0
info:
  description: "Online task planning system — manage tasks, work slots, dependencies\
    \ and generate optimized plans."
  title: ChronoScope API
  version: 1.1.0
servers:
- description: Production
  url: https://chronoscope.ni0.team/api
security:
- oauth2: []
...
paths:
  /v1/identity:
    get:
      description: "Retrieve information about the authenticated identity, including\
        \ all linked accounts and the organizations for which the identity has admin\
        \ privileges, as resolved server-side."
      operationId: getIdentity
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdentityResponse"
          description: Identity retrieved successfully
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
      summary: Get current identity
      tags:
      - Identity
  /v1/identity/accounts:
    post:
      description: "Send a confirmation link to the target e-mail address. If the\
        \ target accepts, the two accounts' identities are merged."
      operationId: requestAccountLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountLinkRequest"
        required: true
      responses:
        "202":
          description: Link e-mail sent
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation error
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
      summary: Request account linking
      tags:
      - Identity
  /v1/identity/accounts/confirm:
    post:
      description: Confirm a pending account link using the token received via e-mail.
        Returns the IDs of both linked accounts.
      operationId: confirmAccountLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountLinkConfirmRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountLinkConfirmResponse"
          description: Account link confirmed
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation error or invalid/expired token
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
      summary: Confirm account linking
      tags:
      - Identity
...
components:
  schemas:
    AccountLinkConfirmRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    AccountLinkConfirmResponse:
      type: object
      properties:
        sourceAccountId:
          type: integer
          format: int64
        status:
          type: string
        targetAccountId:
          type: integer
          format: int64
      required:
      - sourceAccountId
      - status
      - targetAccountId
    AccountLinkRequest:
      type: object
      properties:
        targetEmail:
          type: string
          format: email
          minLength: 1
      required:
      - targetEmail
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        identityId:
          type: integer
          format: int64
      required:
      - id
      - identityId
...
    IdentityResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        adminOrganizations:
          type: array
          description: "Organization IDs for which the authenticated identity has\
            \ admin privileges, resolved server-side via Keycloak Admin Client lookups."
          items:
            type: string
          uniqueItems: true
        id:
          type: integer
          format: int64
        organizations:
          type: array
          description: Organizations to which the identity belongs.
          items:
            $ref: "#/components/schemas/Organization"
          uniqueItems: true
      required:
      - accounts
      - adminOrganizations
      - id
      - organizations
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...
  securitySchemes:
    oauth2:
      description: Keycloak OAuth2 (Authorization Code flow)
      flows:
        authorizationCode:
          authorizationUrl: https://auth.ni0.team/realms/ni0/protocol/openid-connect/auth
          scopes:
            read: Read access
            write: Write access
          tokenUrl: https://auth.ni0.team/realms/ni0/protocol/openid-connect/token
      type: oauth2