# src/main/java/de/ni0/chronoscope/controller/IdentityController.java
package de.ni0.chronoscope.controller;

import de.ni0.chronoscope.model.Identity;
import de.ni0.chronoscope.service.KeycloakService;
import org.keycloak.representations.idm.OrganizationRepresentation;
import org.springframework.http.HttpStatus;
import org.springframework.http.ProblemDetail;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import de.ni0.chronoscope.config.RequestContext;
import de.ni0.chronoscope.controller.dto.request.AccountLinkConfirmRequest;
import de.ni0.chronoscope.controller.dto.request.AccountLinkRequest;
import de.ni0.chronoscope.controller.dto.response.AccountLinkConfirmResponse;
import de.ni0.chronoscope.controller.dto.response.IdentityResponse;
import de.ni0.chronoscope.mapper.IdentityMapper;
import de.ni0.chronoscope.service.IdentityService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

import java.util.HashSet;
import java.util.List;
import java.util.Set;

/**
 * REST controller for identity operations.
 *
 * <p>Exposes the current authenticated identity and account linking endpoints.
 */
@Tag(name = "Identity", description = "Retrieve the current identity and manage account linking")
@RestController
@RequestMapping("/v1/identity")
@RequiredArgsConstructor
public class IdentityController {

    private final IdentityService identityService;
    private final RequestContext requestContext;
    private final IdentityMapper identityMapper;
    private final KeycloakService keycloakService;

    /**
     * Retrieves the authenticated identity.
     *
     * @return the identity response including linked accounts
     */
    @Operation(summary = "Get current identity", description = "Retrieve information about the authenticated identity, including all linked accounts and the organizations for which the identity has admin privileges, as resolved server-side.")    @ApiResponses({
        @ApiResponse(responseCode = "200", description = "Identity retrieved successfully"),
        @ApiResponse(responseCode = "401", description = "Missing or invalid token", content = @Content(schema = @Schema(implementation = ProblemDetail.class)))
    })
    @GetMapping
    public IdentityResponse getIdentity() {
        Identity identity = this.identityService.getIdentity(this.requestContext.getAccount().getIdentity().getId());
        Set<String> adminOrganizations = this.keycloakService.getAdminOrganizations(identity);
        Set<OrganizationRepresentation> organizations = this.keycloakService.getIdentityOrganizations(identity);
        List<IdentityResponse.Organization> orgs = new java.util.ArrayList<>(organizations.stream().map(o ->
                new IdentityResponse.Organization(o.getName(), o.getId())).toList());
        orgs.add(new IdentityResponse.Organization("Privat", "private"));
        return this.identityMapper.toResponse(
            identity,
            adminOrganizations,
            new HashSet<>(orgs)
        );
    }

    /**
     * Requests an account link for the authenticated identity.
     *
     * @param request the account link request payload
     */
    @Operation(summary = "Request account linking", description = "Send a confirmation link to the target e-mail address. If the target accepts, the two accounts' identities are merged.")
    @ApiResponses({
        @ApiResponse(responseCode = "202", description = "Link e-mail sent"),
        @ApiResponse(responseCode = "400", description = "Validation error", content = @Content(schema = @Schema(implementation = ProblemDetail.class))),
        @ApiResponse(responseCode = "401", description = "Missing or invalid token", content = @Content(schema = @Schema(implementation = ProblemDetail.class)))
    })
    @PostMapping("/accounts")
    @ResponseStatus(HttpStatus.ACCEPTED)
    public void requestAccountLink(@Valid @RequestBody AccountLinkRequest request) {
        long accountId = this.requestContext.getAccount().getId();
        String targetEmail = request.targetEmail();
        this.identityService.sendLink(accountId, targetEmail);
    }

    /**
     * Confirms an account-link token and merges the target account's identity into the source identity.
     *
     * @param request the confirmation token payload
     * @return IDs of the accounts involved in the merge
     */
    @Operation(summary = "Confirm account linking", description = "Confirm a pending account link using the token received via e-mail. Returns the IDs of both linked accounts.")
    @ApiResponses({
        @ApiResponse(responseCode = "200", description = "Account link confirmed"),
        @ApiResponse(responseCode = "400", description = "Validation error or invalid/expired token", content = @Content(schema = @Schema(implementation = ProblemDetail.class))),
        @ApiResponse(responseCode = "401", description = "Missing or invalid token", content = @Content(schema = @Schema(implementation = ProblemDetail.class)))
    })
    @PostMapping("/accounts/confirm")
    public AccountLinkConfirmResponse confirmAccountLink(@Valid @RequestBody AccountLinkConfirmRequest request) {
        long identityId = this.requestContext.getAccount().getIdentity().getId();
        return this.identityService.mergeAccounts(identityId, request.token());
    }
}